feat(phase3): cross-subject comparator, mock test suite, decode improvements + eval fixes
Hardening pass before results/thesis writing. Dev on RTX 2070 (CPU/mock tests),
recon runs on RTX 4070 via git pull.

- evaluation.py: fix crash on removed config.NSD_CONFIG; match real recon
  filenames ({subject}/reconstructions/{stem}_recon.png) and resolve GT by stem
- phase2/compare_subjects.py: new [GT|CSI1|CSI2|CSI3|CSI4] single-sample figure
  over the shared 113 repeated stimuli (torch-free, runs post git-pull)
- train_mindeye/mindeye_models: opt-in BiMixCo->SoftCLIP schedule (--mixco),
  MindEye pairwise-ID lift; default path unchanged
- visual_evaluator: configurable embedding norm (--embed-norm none for MindEye;
  the hardcoded *12.0 was a Ridge-shrinkage patch)
- mock_data: honor float32 Split contract (was upcasting to float64)
- tests/: 24 mock tests incl. full reconstruction dry-run (no GPU/diffusers)
- docs/fase3.md + docs/RAG_reconstruccion_fmri.md: improvement plan + SOTA digest

+++ src/phase2/mock_data.py
@@ -52,11 +52,12 @@ def make_mock_split(
     clip_train_clean = betas_train @ W_true
     clip_test_clean = betas_test @ W_true
 
-    noise_std_train = clip_train_clean.std() / snr
-    noise_std_test = clip_test_clean.std() / snr
+    # `.std()` devuelve float64; castear el escalar evita upcastear los arrays.
+    noise_std_train = np.float32(clip_train_clean.std() / snr)
+    noise_std_test = np.float32(clip_test_clean.std() / snr)
 
-    clip_train = clip_train_clean + rng.standard_normal(clip_train_clean.shape).astype(np.float32) * noise_std_train
-    clip_test = clip_test_clean + rng.standard_normal(clip_test_clean.shape).astype(np.float32) * noise_std_test
+    clip_train = (clip_train_clean + rng.standard_normal(clip_train_clean.shape).astype(np.float32) * noise_std_train).astype(np.float32)
+    clip_test = (clip_test_clean + rng.standard_normal(clip_test_clean.shape).astype(np.float32) * noise_std_test).astype(np.float32)
 
     trial_ids_train = list(range(n_train))
     trial_ids_test = list(range(n_train, n_train + n_test))
